# contracts/requirement-internal.openapi.json
{
  "components": {
    "schemas": {
...
      "BulkResult": {
        "properties": {
          "requirements": {
            "title": "Requirements",
            "type": "integer"
          }
        },
        "required": [
          "requirements"
        ],
        "title": "BulkResult",
        "type": "object"
      },
      "CorpusActionRequest": {
        "description": "The knowledge admin acting, and why: recorded with the action and told to the owner.",
        "properties": {
          "actor_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "actor_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Name",
            "type": "string"
          },
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "actor_name",
          "reason"
        ],
        "title": "CorpusActionRequest",
        "type": "object"
      },
      "CorpusPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CorpusRow"
            },
            "title": "Items",
            "type": "array"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "required": [
          "items",
          "next_offset"
        ],
        "title": "CorpusPage",
        "type": "object"
      },
      "CorpusRow": {
        "properties": {
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "index_state": {
            "$ref": "#/components/schemas/IndexState"
          },
          "last_screened_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Screened At"
          },
          "open_findings": {
            "title": "Open Findings",
            "type": "integer"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonName"
              },
              {
                "type": "null"
              }
            ]
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "retired": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetiredMark"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "requirement_id",
          "title",
          "duplicate",
          "owner",
          "index_state",
          "last_screened_at",
          "open_findings",
          "retired"
        ],
        "title": "CorpusRow",
        "type": "object"
      },
      "CorpusState": {
        "enum": [
          "active",
          "retired"
        ],
        "title": "CorpusState",
        "type": "string"
      },
...
      "DependencyImpactPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DependencyImpact"
            },
            "title": "Items",
            "type": "array"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "required": [
          "items",
          "next_offset"
        ],
        "title": "DependencyImpactPage",
        "type": "object"
      },
      "DependentsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceDependency"
            },
            "title": "Items",
            "type": "array"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "required": [
          "items",
          "next_offset"
        ],
        "title": "DependentsPage",
        "type": "object"
      },
      "FindingAge": {
        "enum": [
          "under_7_days",
          "from_7_to_30_days",
          "over_30_days"
        ],
        "title": "FindingAge",
        "type": "string"
      },
...
      "FindingsPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FindingRow"
            },
            "title": "Items",
            "type": "array"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          }
        },
        "required": [
          "items",
          "next_offset"
        ],
        "title": "FindingsPage",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IndexState": {
        "enum": [
          "current",
          "waiting",
          "failed",
          "rebuild_required"
        ],
        "title": "IndexState",
        "type": "string"
      },
      "KnowledgeRelationshipKind": {
        "enum": [
          "possible_duplicate",
          "possible_contradiction"
        ],
        "title": "KnowledgeRelationshipKind",
        "type": "string"
      },
...
      "MembershipResult": {
        "properties": {
          "changed_at": {
            "format": "date-time",
            "title": "Changed At",
            "type": "string"
          },
          "closed_findings": {
            "title": "Closed Findings",
            "type": "integer"
          },
          "notified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notified"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/CorpusState"
          }
        },
        "required": [
          "requirement_id",
          "state",
          "changed_at",
          "closed_findings",
          "notified"
        ],
        "title": "MembershipResult",
        "type": "object"
      },
...
      "NudgeRequest": {
        "description": "The knowledge admin who asks: named in the record and in each owner's notification.",
        "properties": {
          "actor_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "actor_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Name",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "actor_name"
        ],
        "title": "NudgeRequest",
        "type": "object"
      },
      "NudgeResult": {
        "properties": {
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "next_nudge_at": {
            "format": "date-time",
            "title": "Next Nudge At",
            "type": "string"
          },
          "nudged_at": {
            "format": "date-time",
            "title": "Nudged At",
            "type": "string"
          },
          "recipients": {
            "items": {
              "type": "string"
            },
            "title": "Recipients",
            "type": "array"
          }
        },
        "required": [
          "finding_id",
          "nudged_at",
          "recipients",
          "next_nudge_at"
        ],
        "title": "NudgeResult",
        "type": "object"
      },
...
      "ReindexRequest": {
        "properties": {
          "actor_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "actor_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Name",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "requirement_ids": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 500,
            "title": "Requirement Ids",
            "type": "array"
          },
          "scope": {
            "enum": [
              "failed",
              "requirements"
            ],
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "actor_name",
          "scope"
        ],
        "title": "ReindexRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Requirement service internal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/knowledge/corpus": {
      "get": {
        "description": "The corpus Requirement by Requirement: identity and state, never content.",
        "operationId": "knowledge_corpus_internal_knowledge_corpus_get",
        "parameters": [
          {
            "in": "query",
            "name": "index_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IndexState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Index State"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "open_findings_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Open Findings Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "not_screened_for_days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 3650,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Not Screened For Days"
            }
          },
          {
            "in": "query",
            "name": "retired_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Retired Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorpusPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Knowledge Corpus",
        "tags": [
          "internal"
        ]
      }
    },
...
    "/internal/knowledge/findings": {
      "get": {
        "description": "Findings in force across the corpus, the longest-standing first.",
        "operationId": "knowledge_findings_internal_knowledge_findings_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/KnowledgeRelationshipKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "age",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FindingAge"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Age"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Knowledge Findings",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/knowledge/findings/{finding_id}/nudge": {
      "post": {
        "description": "Ask both Requirements' owners to decide a finding; at most once a week.",
        "operationId": "nudge_finding_owners_internal_knowledge_findings__finding_id__nudge_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NudgeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NudgeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nudge Finding Owners",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/knowledge/reindex": {
      "post": {
        "description": "Retry what stopped indexing, or index chosen Requirements again; the worker does it.",
        "operationId": "reindex_corpus_internal_knowledge_reindex_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReindexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reindex Corpus",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/knowledge/requirements/{requirement_id}/reinstatement": {
      "post": {
        "description": "Return a retired Requirement to the corpus; it is indexed again and screened afresh.",
        "operationId": "reinstate_to_corpus_internal_knowledge_requirements__requirement_id__reinstatement_post",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorpusActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reinstate To Corpus",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/knowledge/requirements/{requirement_id}/retirement": {
      "post": {
        "description": "Take a Requirement out of the corpus; every open finding citing it closes.",
        "operationId": "retire_from_corpus_internal_knowledge_requirements__requirement_id__retirement_post",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorpusActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire From Corpus",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/references/{document_id}/dependents": {
      "get": {
        "operationId": "document_dependents_internal_references__document_id__dependents_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Actor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Kind"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependentsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Document Dependents",
        "tags": [
          "internal"
        ]
      }
    },
    "/internal/references/{document_id}/impact": {
      "get": {
        "operationId": "document_impact_internal_references__document_id__impact_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor_id",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Actor Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Query",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependencyImpactPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Document Impact",
        "tags": [
          "internal"
        ]
      }
    }
  }
}